requirements.txt
discord
aiohttp
python-dotenv
ffmpeg
TenGiphPy
youtube_dl
asyncio